refactor: share pointInTriangle utility
