+++ systems/drill.js
@@ -1,4 +1,5 @@
 // systems/drill.js
+import { pointInTriangle } from "../utils/geometry.js";
 
 export function initDrill(state, opts = {}) {
   const r = state.player?.r ?? 18;
@@ -151,11 +152,4 @@ export function carveObstaclesWithDrillTri(state, tri, dt, pad = 2) {
   }
 }
 
-// Helper: point-in-triangle test (same math used in enemies.js)
-function pointInTriangle(p, a, b, c) {
-  const areaOrig = Math.abs((b.x - a.x) * (c.y - a.y) - (c.x - a.x) * (b.y - a.y));
-  const area1 = Math.abs((a.x - p.x) * (b.y - p.y) - (b.x - p.x) * (a.y - p.y));
-  const area2 = Math.abs((b.x - p.x) * (c.y - p.y) - (c.x - p.x) * (b.y - p.y));
-  const area3 = Math.abs((c.x - p.x) * (a.y - p.y) - (a.x - p.x) * (c.y - p.y));
-  return Math.abs(area1 + area2 + area3 - areaOrig) < 0.01;
-}
+
