+++ systems/enemies.js
@@ -2,6 +2,7 @@
 import { spawnPickup } from "./pickups.js";
 import { collideWithObstacles } from "./world.js";
 import { getDrillTriangleWorld } from "./drill.js";
+import { pointInTriangle } from "../utils/geometry.js";
 
 export function initEnemies(state, opts = {}) {
   const cols = state.miasma.cols ?? (state.miasma.halfCols * 2);
@@ -335,11 +336,3 @@ function randInt(min, max) {
   return (min + Math.floor(Math.random() * (max - min + 1)));
 }
 
-// Used by drill damage AABB pass
-function pointInTriangle(p, a, b, c) {
-  const areaOrig = Math.abs((b.x - a.x) * (c.y - a.y) - (c.x - a.x) * (b.y - a.y));
-  const area1 = Math.abs((a.x - p.x) * (b.y - p.y) - (b.x - p.x) * (a.y - p.y));
-  const area2 = Math.abs((b.x - p.x) * (c.y - p.y) - (c.x - p.x) * (b.y - p.y));
-  const area3 = Math.abs((c.x - p.x) * (a.y - p.y) - (a.x - p.x) * (c.y - p.y));
-  return Math.abs(area1 + area2 + area3 - areaOrig) < 0.01;
-}
